-changed project to v140 toolset -implemented custom Combo for TravelPanel with keyboard interaction -added /show and /hide hotkeys for panels and windows -fixed content-based chat filter -fixed default theme -added options to hide components of info panel -added timestamps to resign log -reordered and improved pcons settings -added option to close travel panel on travel -notepad now saves and loads content -polished minimap settings

+++ GWToolbox/GWToolbox/GWToolbox.cpp
@@ -94,7 +94,7 @@ void GWToolbox::ThreadEntry(HMODULE dllmodule) {
 	GW::Cameramgr();
 
 	dx_hooker = new GW::DirectXHooker();
-	printf("DxDevice = %X\n", dx_hooker->device());
+	printf("DxDevice = %X\n", (unsigned int)(dx_hooker->device()));
 
 	LOG("Installing dx hooks\n");
 	dx_hooker->AddHook(GW::dx9::kEndScene, (void*)EndScene);
